common: silence GCC warnings
Preforker.h:111:8: warning: ignoring return value of ‘ssize_t
safe_write(int, const void*, size_t)’, declared with attribute
warn_unused_result [-Wunused-result]
        (void)safe_write(fd[1], &r, sizeof(r));
        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

and

compat.cc:36:28: warning: comparison between signed and unsigned integer
expressions [-Wsign-compare]
     if (off + sizeof(data) > len)
         ~~~~~~~~~~~~~~~~~~~^~~~~

Fixes: http://tracker.ceph.com/issues/23774

+++ src/common/compat.cc
@@ -33,7 +33,7 @@ int manual_fallocate(int fd, off_t offset, off_t len) {
   // TODO: compressing filesystems would require random data
   memset(data, 0x42, sizeof(data));
   for (off_t off = 0; off < len; off += sizeof(data)) {
-    if (off + sizeof(data) > len)
+    if (off + static_cast<off_t>(sizeof(data)) > len)
       r = safe_write(fd, data, len - off);
     else
       r = safe_write(fd, data, sizeof(data));
